Created sparate components for the landing page to make the code more clean and scaleable.
