+++ app/components/Footer.tsx
@@ -1,0 +1,9 @@
+export default function Footer() {
+  return (
+    <footer className="bg-gray-900 text-gray-300 py-8 text-center">
+      <p>
+        &copy; {new Date().getFullYear()} The Experience. All rights reserved.
+      </p>
+    </footer>
+  );
+};
