+++ app/globals.css
@@ -1,5 +1,6 @@
 @import "tailwindcss";
-@import "tailwindcss/preflight": root {
+
+:root {
   --background: #ffffff;
   --foreground: #171717;
 }
